Updated the following:
Got the average fare for each city in the urban cities. Got the average fare for each city. Changed the name of the urban_driver_count to urban_driver_count_avg because it is a mean. Columbia should consider editing and revising the code. Got the average number of drivers for each suburban and rural cities. Made the same alteration on the variable name.