Final Commit Readme
Updated the readme to reflect the data analysis for the challenge. In doing so, I realized that this is just the beginning of possibilities for reading the data PyBer collects. There are further studies that I could do with a data analytics team to find deeper insight into the business and how the actions affect it. 🍻